+++ src/components/tag/tag.tsx
@@ -1,50 +1,60 @@
+import React from 'react';
+import { FunctionComponent } from 'react';
 import styled, { css } from 'styled-components';
 
 export interface TagProps {
-  /**
-   * Color?
-   */
+  /* Content of Tag */
+  children?: string;
+
+  /* Color of tag */
   color: string;
-  /**
-   * Size of tag
-   */
+
+  /* Size of tag */
   size?: 'small' | 'medium' | 'large';
 }
 
-export const Tag = styled.div<TagProps>`
+const StyledTag = styled.div<TagProps>`
   border: 1.5px solid ${(props) => props.color};
   border-radius: 4px;
-  padding: 0.35rem 0.5rem;
-  text-transform: uppercase;
   color: ${(props) => props.color};
   background-color: ${(props) => props.color}30;
-  font-family: sans-serif;
   font-weight: 500;
   text-align: center;
   display: inline-block;
   vertical-align: middle;
+  margin: 4px;
   ${(props) =>
     props.size === 'small' &&
     css`
-      font-size: 12px;
+      font-size: 10px;
       padding: 0.15rem 0.3rem;
     `}
   ${(props) =>
     props.size === 'medium' &&
     css`
-      font-size: 14px;
-      padding: 0.25rem 0.4rem;
+      font-size: 12px;
+      padding: 0.15rem 0.4rem;
     `}
     ${(props) =>
     props.size === 'large' &&
     css`
-      font-size: 16px;
+      font-size: 14px;
+      padding: 0.075rem 0.5rem;
     `};
 `;
+
+export const Tag: FunctionComponent<TagProps> = ({ children, color, size }) => {
+  return (
+    <StyledTag color={color} size={size}>
+      {children.toUpperCase()}
+    </StyledTag>
+  );
+};
 
 Tag.defaultProps = {
   color: '#BB65FF',
   size: 'small',
+  children: 'Tag',
 };
 
 Tag.displayName = 'Tag';